@@ -4,10 +4,20 @@ import org.bukkit.entity.Player;
 
 public interface Announceable {
     void announceStart(boolean center);
+
     void sendStart(Player player);
+
     void announceEnd();
+
     void announceWinner(boolean center, Player player);
-    void showActionbar();
+
+    void showBossBar();
+
+    void stopBossBar();
+
+    void showActionBar();
+
     void stopActionBar();
+
     String replacePlaceholders(String message);
 }
