@@ -19,8 +19,8 @@ import be.woutzah.chatbrawl.util.ErrorHandler;
 import be.woutzah.chatbrawl.util.FireWorkUtil;
 import be.woutzah.chatbrawl.util.Printer;
 import com.meowj.langutils.lang.LanguageHelper;
-import net.md_5.bungee.api.ChatMessageType;
-import net.md_5.bungee.api.chat.TextComponent;
+import net.kyori.adventure.text.Component;
+import net.kyori.adventure.text.serializer.legacy.LegacyComponentSerializer;
 import org.bukkit.Bukkit;
 import org.bukkit.GameMode;
 import org.bukkit.Material;
@@ -103,7 +103,7 @@ public class FishRace extends ContestantRace {
             if (player.getGameMode() == GameMode.CREATIVE) return;
         }
         World world = player.getWorld();
-        if (!raceManager.isWorldAllowed(world.toString())) return;
+        if (!raceManager.isWorldAllowed(world.getName())) return;
         if (e.getState().equals(PlayerFishEvent.State.CAUGHT_FISH)) {
             Item caughtItem = (Item) e.getCaught();
             if (caughtItem == null) return;
@@ -145,14 +145,17 @@ public class FishRace extends ContestantRace {
     }
 
     @Override
-    public void showActionbar() {
-        String message = replacePlaceholders(settingManager.getString(RaceType.FISH, RaceSetting.LANGUAGE_ACTIONBAR));
+    public void showBossBar() {
+
+    }
+
+    @Override
+    public void showActionBar() {
+        Component message = LegacyComponentSerializer.legacyAmpersand().deserialize(replacePlaceholders(settingManager.getString(RaceType.FISH, RaceSetting.LANGUAGE_ACTIONBAR)));
         this.actionBarTask = new BukkitRunnable() {
             @Override
             public void run() {
-                Bukkit.getOnlinePlayers().forEach(p -> p.spigot()
-                        .sendMessage(ChatMessageType.ACTION_BAR,
-                                new TextComponent(Printer.parseColor(message))));
+                Bukkit.getServer().sendActionBar(message);
             }
         }.runTaskTimer(ChatBrawl.getInstance(), 0, 20);
     }
@@ -170,6 +173,6 @@ public class FishRace extends ContestantRace {
         super.beforeRaceStart();
         initRandomFishEntry();
         if (isAnnounceStartEnabled()) announceStart(isCenterMessages());
-        if (isActionBarEnabled()) showActionbar();
+        if (isActionBarEnabled()) showActionBar();
     }
 }
